+++ packages/account-communicator/src/index.ts
@@ -1,0 +1,3 @@
+export * from './communicator';
+export * from './errors';
+export * from './popup-communicator';