+++ packages/account-react-native/src/context/sophon-context.tsx
@@ -5,6 +5,7 @@ import {
   type DataScopes,
   type SophonNetworkType,
 } from '@sophon-labs/account-core';
+import type { EIP1193Provider } from '@sophon-labs/account-provider';
 import { createContext, useCallback, useMemo, useState } from 'react';
 import {
   type Address,
@@ -15,7 +16,7 @@ import {
 } from 'viem';
 import { sophon, sophonTestnet } from 'viem/chains';
 import { erc7846Actions } from 'viem/experimental';
-import type { WalletProvider } from 'zksync-sso';
+import { eip712WalletActions } from 'viem/zksync';
 import type { SophonJWTToken } from '@/types';
 import { SophonMainView, type SophonMainViewProps } from '../components';
 import { useUIEventHandler } from '../messaging';
@@ -32,7 +33,7 @@ export interface SophonContextConfig {
   account?: SophonAccount;
   setAccount: (account?: SophonAccount) => void;
   chain: Chain;
-  provider?: WalletProvider;
+  provider?: EIP1193Provider;
   network: SophonNetworkType;
   accessToken?: SophonJWTToken | null;
   refreshToken?: SophonJWTToken | null;
@@ -147,7 +148,9 @@ export const SophonContextProvider = ({
   const walletClient = createWalletClient({
     chain: chain,
     transport: custom(provider),
-  }).extend(erc7846Actions());
+  })
+    .extend(erc7846Actions())
+    .extend(eip712WalletActions());
 
   const disconnect = useCallback(() => {
     provider?.disconnect();