+++ packages/account-message-bridge/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @sophon-labs/account-message-bridge
 
+## 2.0.10-next.9
+
+### Patch Changes
+
+- Error checking on react native
+
+## 2.0.10-next.8
+
+### Patch Changes
+
+- Adding addEventListener check
+
+## 2.0.10-next.7
+
+### Patch Changes
+
+- Fixing problem with wagmi chained calls
+
+## 2.0.10-next.6
+
+### Patch Changes
+
+- Correctly considering ssr
+
+## 2.0.10-next.5
+
+### Patch Changes
+
+- Fixing provider window error on react native
+
+## 2.0.10-next.4
+
+### Patch Changes
+
+- Fixing error on react native
+
+## 2.0.10-next.3
+
+### Patch Changes
+
+- Removing zksync-sso from core packages
+
+## 2.0.10-next.2
+
+### Patch Changes
+
+- Simplifying connector and provider dependencies
+
+## 2.0.10-next.1
+
+### Patch Changes
+
+- Adding webauthn as direct dependency to connector package
+
+## 2.0.10-next.0
+
+### Patch Changes
+
+- Exposing sophon connector and sophon provider helper utility
+
 ## 2.0.9
 
 ### Patch Changes